Collect custom modifiers to single file
