--- grid/src/commonMain/kotlin/com/cheonjaeung/compose/grid/SequentialGridAlignmentModifier.kt
@@ -1,46 +1,0 @@
-package com.cheonjaeung.compose.grid
-
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.node.ModifierNodeElement
-import androidx.compose.ui.node.ParentDataModifierNode
-import androidx.compose.ui.platform.InspectorInfo
-import androidx.compose.ui.unit.Density
-
-internal class SequentialGridAlignmentElement(
-    val alignment: Alignment
-) : ModifierNodeElement<SequentialGridAlignmentNode>() {
-    override fun create(): SequentialGridAlignmentNode {
-        return SequentialGridAlignmentNode(alignment)
-    }
-
-    override fun update(node: SequentialGridAlignmentNode) {
-        node.alignment = alignment
-    }
-
-    override fun InspectorInfo.inspectableProperties() {
-        name = "align"
-        value = alignment
-    }
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null) return false
-        if (other !is SequentialGridAlignmentElement) return false
-        return this.alignment == other.alignment
-    }
-
-    override fun hashCode(): Int {
-        return alignment.hashCode()
-    }
-}
-
-internal class SequentialGridAlignmentNode(
-    var alignment: Alignment
-) : Modifier.Node(), ParentDataModifierNode {
-    override fun Density.modifyParentData(parentData: Any?): Any {
-        val p = parentData as? SequentialGridParentData ?: SequentialGridParentData()
-        p.alignment = alignment
-        return p
-    }
-}
